+++ apps/deploy-web/src/components/new-deployment/NewDeploymentContainer.tsx
@@ -3,6 +3,7 @@ import { FC, useEffect, useState } from "react";
 import { useAtomValue } from "jotai";
 import { useRouter, useSearchParams } from "next/navigation";
 
+import { USER_TEMPLATE_CODE } from "@src/config/deploy.config";
 import { CI_CD_TEMPLATE_ID } from "@src/config/remote-deploy.config";
 import { useLocalNotes } from "@src/context/LocalNoteProvider";
 import { useSdlBuilder } from "@src/context/SdlBuilderProvider";
@@ -55,41 +56,41 @@ export const NewDeploymentContainer: FC = () => {
         })
       );
     } else {
-      if (isGitProvider) {
-        setIsGitProviderTemplate(true);
-      } else {
-        setIsGitProviderTemplate(false);
-      }
+      setIsGitProviderTemplate(!!isGitProvider);
     }
   }, [searchParams]);
 
   useEffect(() => {
     const templateId = searchParams?.get("templateId");
     const isCreating = !!activeStep && activeStep > getStepIndexByParam(RouteStep.chooseTemplate);
+
     if (!templates || (isCreating && !!editedManifest && !!templateId)) return;
 
-    const template = getRedeployTemplate() || getGalleryTemplate();
+    const template = getRedeployTemplate() || getGalleryTemplate() || deploySdl;
+    const isUserTemplate = template?.code === USER_TEMPLATE_CODE;
+    const isUserTemplateInit = isUserTemplate && !!editedManifest;
+    if (!template || isUserTemplateInit) return;
 
-    if (template) {
-      setSelectedTemplate(template as TemplateCreation);
-      setEditedManifest(template.content as string);
+    setSelectedTemplate(template as TemplateCreation);
+    setEditedManifest(template.content as string);
 
-      if ("config" in template && (template.config?.ssh || (!template.config?.ssh && hasComponent("ssh")))) {
-        toggleCmp("ssh");
+    if ("config" in template && (template.config?.ssh || (!template.config?.ssh && hasComponent("ssh")))) {
+      toggleCmp("ssh");
+    }
+
+    const cicdTemplate = getTemplateById(CI_CD_TEMPLATE_ID);
+    const isRemoteYamlImage = isImageInYaml(template?.content as string, cicdTemplate?.deploy);
+    const queryStep = searchParams?.get("step");
+    if (queryStep !== RouteStep.editDeployment) {
+      if (isRemoteYamlImage) {
+        setIsGitProviderTemplate(true);
       }
-      const isRemoteYamlImage = isImageInYaml(template?.content as string, getTemplateById(CI_CD_TEMPLATE_ID)?.deploy);
-      const queryStep = searchParams?.get("step");
-      if (queryStep !== RouteStep.editDeployment) {
-        if (isRemoteYamlImage) {
-          setIsGitProviderTemplate(true);
-        }
 
-        const newParams = isRemoteYamlImage
-          ? { ...searchParams, step: RouteStep.editDeployment, gitProvider: "github" }
-          : { ...searchParams, step: RouteStep.editDeployment };
+      const newParams = isRemoteYamlImage
+        ? { ...searchParams, step: RouteStep.editDeployment, gitProvider: "github" }
+        : { ...searchParams, step: RouteStep.editDeployment };
 
-        router.replace(UrlService.newDeployment(newParams));
-      }
+      router.replace(UrlService.newDeployment(newParams));
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [templates, editedManifest, searchParams, router, toggleCmp, hasComponent, activeStep]);
@@ -136,11 +137,6 @@ export const NewDeploymentContainer: FC = () => {
       if (hardCodedTemplate) {
         return hardCodedTemplate;
       }
-    }
-
-    // Jotai state template
-    if (deploySdl) {
-      return deploySdl;
     }
 
     return null;